Refactored the modules __init__ file.
